% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getMetadata.r
\name{getMetadata}
\alias{getMetadata}
\title{Retrieve Landsat image metadata}
\usage{
getMetadata(imgFile)
}
\arguments{
\item{imgFile}{filename (character). Full path to a Landast LPGS MSS file that includes the
original image ID as the first block of characters.}
}
\value{
data.frame with image information.
}
\description{
Uses the image file name to find the corresponding *MTL.txt image metadata file provided with LPSG Landsat images
and returns a dataframe with image information
}
\examples{
\dontrun{

getMetadata("C:/mss/LM10360321973191AAA04/LM10360321973191AAA04_dn.tif")
}
}

