% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MSScvm.r
\name{MSScvm}
\alias{MSScvm}
\title{Landsat MSS cloud and cloud shadow masking}
\usage{
MSScvm(imgDir, demFile, classify = F)
}
\arguments{
\item{imgDir}{directory name (character). Full path to a MSS image directory produced by the \code{\link{MSSunpack}} function.}

\item{demFile}{filename (character). Full path to image-corresponding DEM file.}

\item{classify}{logical. If TRUE clouds, cloud shadows, and clear pixels have unique values (0 = clear, 1 = cloud shadow, 2 = cloud).
If FALSE obscured pixles = 0 and clear = 1.}
}
\value{
A GeoTIFF raster image file with the same dimensions as the MSS image. The file will be placed in the
'imgDir' directory with the name equal to the image ID followed by '_msscvm'.
}
\description{
Creates a cloud and cloud shadow mask for Landsat MSS imagery.
}
\details{
It is important that the input DEM file, specified by the 'demFile' parameter, be the same projection and pixel resolution as the
input image. It must also be >= in spatial extent, relative to the image. The program will check for these attributes and throw an error message if
there is a violation. There are two helper functions to prepare a suitable DEM. Use the \code{\link{reprojectDEM}} function to ensure proper projection
and pixel resolution of an exisiting DEM, and the \code{\link{mosaicDEMs}} function to create a mosaic from several DEMs to ensure proper extent, projection,
and pixel resolution.
}
\examples{
\dontrun{

MSScvm(imgDir = "C:/mss/LM10360321973191AAA04",
       demFile = "C:/mss/dem/wrs1_p036r032_dem.tif")
}
}

