% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MSSdn2rad.r
\name{MSSdn2rad}
\alias{MSSdn2rad}
\title{Convert MSS DN values to TOA radiance}
\usage{
MSSdn2rad(imgFile)
}
\arguments{
\item{imgFile}{filename (character). Full path to *dn.tif image file produced by the \code{\link{MSSunpack}} function.}
}
\value{
A 4-band Landsat MSS GeoTIFF raster image file in units of top-of-atmosphere (TOA) radiance. The file will be placed in
same directory as the 'imgFile' with the name equal to the image ID followed by 'toa_radiance'. Note that the values are scaled by 100
and rounded to the nearest integer to reduce the file size.
}
\description{
Convert MSS DN values to TOA radiance
}
\details{
The equation used to convert DN to TOA radiance can be found \href{http://landsat.usgs.gov/how_is_radiance_calculated.php}{here}.
}
\examples{
\dontrun{

MSSdn2rad("C:/mss/LM10360321973191AAA04/LM10360321973191AAA04_dn.tif")
}
}
\seealso{
\code{\link{MSSdn2refl}}
}

