% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MSScvm.r
\name{MSScvm}
\alias{MSScvm}
\title{Create a cloud and cloud shadow mask for Landsat MSS imagery}
\usage{
MSScvm(imgDir, demFile, classify = F)
}
\arguments{
\item{imgDir}{directory name (character). Full path to a MSS image directory produced by the \code{\link{MSSunpack}} function.}

\item{demFile}{filename (character). Full path to image-corresponding DEM file.}

\item{classify}{logical. If TRUE clouds, cloud shadows, and clear pixels have unique values. If FALSE obscured pixels are value 0 and clear are 1.}
}
\value{
A binary raster with the same dimensions as the MSS image where pixels with value 1 represent clear pixel and 0 as obsured by either cloud or cloud shadow
}
\description{
Create a cloud and cloud shadow mask for Landsat MSS imagery
}
\details{
bla
}
\examples{
\dontrun{

MSScvm(imgDir = "C:/mss/LM10360321973191AAA04",
       demFile = "C:/mss/dem/wrs1_p036r032_dem.tif")
}
}

